+++ source/extensions/isaacsim.robot.manipulators.ui/isaacsim/robot/manipulators/ui/tests/test_menu_graphs.py
@@ -199,7 +199,7 @@ class TestArticulationGraphs(OmniUiTest):
         if assets_root_path is None:
             carb.log_error("Could not find Isaac Sim assets folder")
             return
-        robot_prim.GetReferences().AddReference(assets_root_path + "/Isaac/Robots/Jetbot/jetbot.usd")
+        robot_prim.GetReferences().AddReference(assets_root_path + "/Isaac/Robots/NVIDIA/Jetbot/jetbot.usd")
         await update_stage_async()
 
         # Create velocity controller graph
@@ -247,7 +247,7 @@ class TestArticulationGraphs(OmniUiTest):
         if assets_root_path is None:
             carb.log_error("Could not find Isaac Sim assets folder")
             return
-        robot_prim.GetReferences().AddReference(assets_root_path + "/Isaac/Robots/Jetbot/jetbot.usd")
+        robot_prim.GetReferences().AddReference(assets_root_path + "/Isaac/Robots/NVIDIA/Jetbot/jetbot.usd")
         # add a ground plane
         await update_stage_async()