+++ source/standalone_examples/api/isaacsim.sensors.rtx/rotating_lidar_rtx.py
@@ -37,7 +37,7 @@ if assets_root_path is None:
     carb.log_error("Could not find Isaac Sim assets folder")
     simulation_app.close()
     sys.exit()
-asset_path = assets_root_path + "/Isaac/Robots/NVIDIA/Carter/nova_carter/nova_carter.usd"
+asset_path = assets_root_path + "/Isaac/Robots/NVIDIA/NovaCarter/nova_carter.usd"
 my_carter = my_world.scene.add(
     WheeledRobot(
         prim_path="/World/Carter",