+++ source/extensions/isaacsim.util.internal/docs/CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
-## [2.1.3] - 2025-04-04
+## [2.1.5] - 2025-04-09
+### Changed
+- Update Isaac Sim NVIDIA robot asset path
+
+## [2.1.4] - 2025-04-04
 ### Changed
 - Version bump to fix extension publishing issues
 
-## [2.1.2] - 2025-03-26
+## [2.1.3] - 2025-03-26
 ### Changed
 - Cleanup and standardize extension.toml, update code formatting for all code